Revision 569: Fix: -npe onMagicFinalizer. -FrequentSkill npe. -Cyrillic characters support in cboard ,thanks to Fortitude for report. -255 tutorial message,thanks to Fortitude for report. -cboard switch typo,thanks to Fortitude for report. -multiply statement , thanks to Fortitude for report. -allow to interact with dead monsters to use "sweep" correctly. -potions are now visible under buffs(l2off like). -hp-mp negative value. -optimize-fix updateAbnormalEffect unnecessary packet broadcast. now it will send the update only if abnormal effects are in use or effected by somone,which is lead to a huge broadcast optimize.
Monster behavior: 
-onAggression remove minion assist(l2 off like),  
-onAggression minions are following master and attacking only when we
attack master-minion.tested on advext.

Skills fix:
-lure(skill) behavior on monster,thanks nijota for report.

Rework:
CharEffectList.java

optimize:
(Custom)EventEgnine: "asynchronize" teleport to avoid massive lag.

+++ L2JHellasC/java/com/l2jhellas/gameserver/handlers/admincommandhandlers/AdminKill.java
@@ -47,10 +47,13 @@ public class AdminKill implements IAdminCommandHandler
 
 						L2World.getInstance().forEachVisibleObjectInRange(player, L2Character.class, radius, knownChar ->
 						{
-							if (knownChar.equals(activeChar))
-								return;
+							if(knownChar != null)
+							{
+								if (knownChar.equals(activeChar))
+									return;
 
-							kill(activeChar, knownChar);
+								kill(activeChar, knownChar);
+							}
 						});
 						activeChar.sendMessage("Killed all characters within a "+ radius + " unit radius around "+ player.getName() + ".");
 					}
@@ -66,7 +69,8 @@ public class AdminKill implements IAdminCommandHandler
 
 				L2World.getInstance().forEachVisibleObjectInRange(activeChar,L2Character.class, radius, knownChar ->
 				{
-					kill(activeChar, knownChar);
+					if(knownChar != null)
+					   kill(activeChar, knownChar);
 				});
 
 				activeChar.sendMessage("Killed all characters within a "+ radius + " unit radius.");
